% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_range.R
\name{check_date_range}
\alias{check_date_range}
\title{check_date_range}
\usage{
check_date_range(data, interval = 8, error = FALSE)
}
\arguments{
\item{data}{data.frame, data. Must contain a \code{date} column.}

\item{interval}{integer, time interval between datasets in hours}

\item{error}{boolean, pass TRUE to raise an error for missing dates}
}
\value{
tibble
}
\description{
Check that all dates are present in a dataset given a certain interval
}
