% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid_reference.R
\name{get_centroid_reference}
\alias{get_centroid_reference}
\title{get_centroid_reference}
\usage{
get_centroid_reference(mob, sf = FALSE)
}
\arguments{
\item{mob}{data.frame, Combined mobility data.}

\item{sf}{boolean, Return an sf object? Defaults to FALSE}
}
\value{
tibble
}
\description{
Get a reference of all tile centroids in a mobility dataset
}
