% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journey_lines.R
\name{get_journey_lines}
\alias{get_journey_lines}
\title{get_journey_lines}
\usage{
get_journey_lines(mob)
}
\arguments{
\item{mob}{data.frame, Combined mobility data.}
}
\value{
tibble
}
\description{
Get a spatial reference of all joureys in a mobility dataset
}
