% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mob.R
\name{read_mob}
\alias{read_mob}
\title{read_mob}
\usage{
read_mob(path, col_types = NULL, zoom_level = 12)
}
\arguments{
\item{path}{string, path to mobility file}

\item{col_types}{function, column data type mapping. Defaults to dplyr::cols}

\item{zoom_level}{integer, dataset zoom level. Defaults to 12}
}
\value{
tibble
}
\description{
Read csv file of mobility data
}
